Fix implicit conversion bug with easy_widebeam().
* Updated full_fov_comparison to have a third slice, widebeam_2tx, and
  to have scanbound for all slices.
* Verified correct operation of full_fov_comparison, bistatic_test, and
  full_fov_15km with oscilloscope and lab system.
* Use sequence_print in sequences.py to give more information about
  antenna phases/magnitudes for each slice.

+++ experiments/superdarn_common_fields.py
@@ -163,7 +163,7 @@ def easy_widebeam(frequency_khz, tx_antennas, antenna_spacing_m):
         13200: [0., 43.42908842, 84.21675093, 174.68458927, 174.68458927, 84.21675093, 43.42908842, 0.]
     }
     num_antennas = opts.main_antenna_count
-    phases = np.zeros(num_antennas)
+    phases = np.zeros(num_antennas, dtype=np.complex64)
     if len(tx_antennas) == 16:
         if frequency_khz in cached_values_16_antennas.keys():
             phases[tx_antennas] = np.exp(1j * np.pi/180. * np.array(cached_values_16_antennas[frequency_khz]))
